discovery/requirements.txt
httpx==0.27.2
psycopg[binary]==3.2.1
pydantic==2.8.2
PyYAML==6.0.2
